requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
django-heroku==0.3.1
django-mysql==3.3.0
django-phone-field==1.8.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
djangorestframework==3.11.0
six==1.14.0
django-filter
django-filter
mapbox